requirements.txt
ultralytics
opencv-python
PyYAML
playsound==1.2.2
simpleaudio